+++ grid-parts/customizer/header-search.php
@@ -12,13 +12,22 @@ Customizer::add_outer_section(
 	'gridd_grid_part_details_header_search',
 	[
 		/* translators: The grid-part label. */
-		'title'       => sprintf( esc_html__( '%s Options', 'gridd' ), esc_html__( 'Header Search', 'gridd' ) ),
-		'description' => Customizer::section_description(
-			'gridd_grid_part_details_header_search',
-			[
-				'docs' => 'https://wplemon.github.io/gridd/grid-parts/search.html',
-			]
-		),
+		'title' => sprintf( esc_html__( '%s Options', 'gridd' ), esc_html__( 'Header Search', 'gridd' ) ),
+	]
+);
+
+Customizer::add_field(
+	[
+		'type'        => 'radio-buttonset',
+		'settings'    => 'gridd_grid_part_details_header_search_mode',
+		'label'       => esc_html__( 'Search Mode', 'gridd' ),
+		'section'     => 'gridd_grid_part_details_header_search',
+		'default'     => 'form',
+		'transport'   => 'refresh',
+		'choices'     => [
+			'form'  => esc_html__( 'Form', 'gridd' ),
+			'slide-up' => esc_html__( 'Slide Up', 'gridd' ),
+		],
 	]
 );