+++ grid-parts/customizer/navigation.php
@@ -44,23 +44,11 @@ function gridd_nav_customizer_options( $id ) {
 	Customizer::add_outer_section(
 		"gridd_grid_part_details_nav_$id",
 		[
-			'title'       => sprintf(
+			'title' => sprintf(
 				/* translators: The grid-part label. */
 				esc_html__( '%s Options', 'gridd' ),
 				/* translators: The navigation number. */
 				sprintf( esc_html__( 'Navigation %d', 'gridd' ), absint( $id ) )
-			),
-			'description' => Customizer::section_description(
-				"gridd_grid_part_details_nav_$id",
-				[
-					'plus' => [
-						esc_html__( 'More extensive control over colors', 'gridd' ),
-						esc_html__( 'WooCommerce Cart inside the navigation', 'gridd' ),
-						esc_html__( 'font-size adjustments', 'gridd' ),
-					],
-					'tip'  => '<p>' . esc_html__( 'Try to keep your horizontal navigation as short as possible and avoid hierarchical menus.', 'gridd' ) . '</p><p>' . esc_html__( 'Verical menus provide a better user experience for hierarchical (nested) menus.', 'gridd' ) . '</p>',
-					'docs' => 'https://wplemon.github.io/gridd/grid-parts/navigation.html',
-				]
 			),
 		]
 	);
@@ -117,8 +105,12 @@ function gridd_nav_customizer_options( $id ) {
 			'description' => Customizer::get_control_description(
 				[
 					'short'   => '',
-					'details' => esc_html__( 'Use any valid CSS value. For details on how padding works, please refer to <a href="https://developer.mozilla.org/en-US/docs/Web/CSS/padding" target="_blank" rel="nofollow">this article</a>.', 'gridd' ),
-				]
+					'details' => sprintf(
+						/* translators: Link properties. */
+						__( 'Use any valid CSS value. For details on how padding works, please refer to <a %s>this article</a>.', 'gridd' ),
+						'href="https://developer.mozilla.org/en-US/docs/Web/CSS/padding" target="_blank" rel="nofollow"'
+					),
+					]
 			),
 			'section'     => "gridd_grid_part_details_nav_$id",
 			'default'     => '1em',